Add only .po for changed .rst files on pull-translations
When translation is ready Crowdin tends to download all
the translated .po files. It can be annoying when the
Crowdin branch has more changed files then it needed
for the current PR. It forces the author to clean PR manually.

The fixed pull-translation workflow contains a filter
which allows to add only corresponding .po files to
the changed .rst files in the current branch.

Closes #2434