ci: remove wakeup request
Request on WAKEUP_URL was added in the time when we deployed to
Heroku, and a server on Heroku used to 'sleep' sometimes, failing to
receive deployment webhooks. After such request it came to a working
state. This request was initially added in #1978.

(cherry picked from commit 02698afd15f0e1b746c245b6ec420972e5e63b06)